Phase A: Create Components module
Extract shared foundation from OCP:
- Create src/Components/ manifest with aliases, TimeDependence, abstract types
- Move shared components (TimesModel, StateModel, etc.) to Components
- Move default functions (__*) to Components
- Define generic functions (state, control, state_dimension, etc.) in Components
- Update OCP to use Components as dependency
- Maintain backward compatibility via reexports in CTModels.jl

Part of OCP split refactor (Components/Models/Building/Solutions)
