+++ reports/notes.md
@@ -1,6 +1,5 @@
 12 juin 2026
 
-- split model and solution
-- split model and premodel?
+- split model and solution and premodel
 - replace closure by callable struct
 - when Traits is in CTBase, use it correctly here.
