+++ src/OCP/Components/control.jl
@@ -89,164 +89,5 @@ function control!(
     return nothing
 end
 
-# ------------------------------------------------------------------------------ #
-# GETTERS
-# ------------------------------------------------------------------------------ #
-"""
-$(TYPEDSIGNATURES)
-
-Get the name of the control variable.
-
-# Arguments
-- `model::ControlModel`: The control model.
-
-# Returns
-- `String`: The name of the control.
-
-# Example
-```julia-repl
-julia> name(controlmodel)
-"u"
-```
-"""
-function name(model::ControlModel)::String
-    return model.name
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Get the name of the control variable from the solution.
-
-# Arguments
-- `model::ControlModelSolution`: The control model solution.
-
-# Returns
-- `String`: The name of the control.
-"""
-function name(model::ControlModelSolution)::String
-    return model.name
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Get the names of the control components.
-
-# Arguments
-- `model::ControlModel`: The control model.
-
-# Returns
-- `Vector{String}`: A list of control component names.
-
-# Example
-```julia-repl
-julia> components(controlmodel)
-["u₁", "u₂"]
-```
-"""
-function components(model::ControlModel)::Vector{String}
-    return model.components
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Get the names of the control components from the solution.
-
-# Arguments
-- `model::ControlModelSolution`: The control model solution.
-
-# Returns
-- `Vector{String}`: A list of control component names.
-"""
-function components(model::ControlModelSolution)::Vector{String}
-    return model.components
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Get the control input dimension.
-
-# Arguments
-- `model::ControlModel`: The control model.
-
-# Returns
-- `Dimension`: The number of control components.
-"""
-function dimension(model::ControlModel)::Dimension
-    return length(components(model))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Get the control input dimension from the solution.
-
-# Arguments
-- `model::ControlModelSolution`: The control model solution.
-
-# Returns
-- `Dimension`: The number of control components.
-"""
-function dimension(model::ControlModelSolution)::Dimension
-    return length(components(model))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Get the control function associated with the solution.
-
-# Arguments
-- `model::ControlModelSolution{TS}`: The control model solution.
-
-# Returns
-- `TS`: A function giving the control value at a given time or state.
-"""
-function value(model::ControlModelSolution{TS})::TS where {TS<:Function}
-    return model.value
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Get the interpolation type for the control.
-
-# Arguments
-- `model::ControlModelSolution`: The control model solution.
-
-# Returns
-- `Symbol`: The interpolation type (`:constant` or `:linear`).
-"""
-function interpolation(model::ControlModelSolution)::Symbol
-    return model.interpolation
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return an empty string, since no control is defined.
-"""
-function name(::EmptyControlModel)::String
-    return ""
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return an empty vector since there are no control components defined.
-"""
-function components(::EmptyControlModel)::Vector{String}
-    return String[]
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return `0` since no control is defined.
-"""
-function dimension(::EmptyControlModel)::Dimension
-    return 0
-end
+# Getters for ControlModel/ControlModelSolution/EmptyControlModel are now in
+# src/Components/accessors.jl.
