+++ src/OCP/Components/variable.jl
@@ -82,96 +82,5 @@ function variable!(
     return nothing
 end
 
-# ------------------------------------------------------------------------------ #
-# GETTERS
-# ------------------------------------------------------------------------------ #
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the variable stored in the model.
-"""
-function name(model::VariableModel)::String
-    return model.name
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the name of the variable stored in the model solution.
-"""
-function name(model::VariableModelSolution)::String
-    return model.name
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return an empty string, since no variable is defined.
-"""
-function name(::EmptyVariableModel)::String
-    return ""
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the names of the components of the variable.
-"""
-function components(model::VariableModel)::Vector{String}
-    return model.components
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the names of the components from the variable solution.
-"""
-function components(model::VariableModelSolution)::Vector{String}
-    return model.components
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return an empty vector since there are no variable components defined.
-"""
-function components(::EmptyVariableModel)::Vector{String}
-    return String[]
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the dimension (number of components) of the variable.
-"""
-function dimension(model::VariableModel)::Dimension
-    return length(components(model))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the number of components in the variable solution.
-"""
-function dimension(model::VariableModelSolution)::Dimension
-    return length(components(model))
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return `0` since no variable is defined.
-"""
-function dimension(::EmptyVariableModel)::Dimension
-    return 0
-end
-
-"""
-$(TYPEDSIGNATURES)
-
-Return the value stored in the variable solution model.
-"""
-function value(model::VariableModelSolution{TS})::TS where {TS<:Union{ctNumber,ctVector}}
-    return model.value
-end
+# Getters for VariableModel/VariableModelSolution/EmptyVariableModel are now in
+# src/Components/accessors.jl.
